sweep: #7656 Deprecate ARC and ARC6
