+++ src/DIRAC/WorkloadManagementSystem/scripts/dirac_jobexec.py
@@ -95,14 +95,12 @@ def main():
     if not jobExec["OK"]:
         gLogger.debug("Workflow execution finished with errors, exiting")
         if jobExec["Errno"]:
-            os._exit(jobExec["Errno"])
+            sys.exit(jobExec["Errno"])
         else:
-            os._exit(1)
+            sys.exit(1)
     else:
         gLogger.debug("Workflow execution successful, exiting")
-        # dirac_jobexec might interact with ARC library which cannot be closed using a simple sys.exit(0)
-        # See https://bugzilla.nordugrid.org/show_bug.cgi?id=4022 for further details
-        os._exit(0)
+        sys.exit(0)
 
 
 if __name__ == "__main__":
